+++ src/components/WeightedResultBar/index.ts
@@ -1,2 +1,1 @@
 export { WeightedResultBar } from './WeightedResultBar';
-export { WeightedResultBarLoader } from './WeightedResultBarLoader';