+++ stories/LocationSelector.tsx
@@ -4,6 +4,7 @@ import { number, text, select, boolean, withKnobs } from '@storybook/addon-knobs
 import { LocationSelector } from '@textkernel/oneui';
 import { ensureApiKey } from './utils/ensureApiKey';
 import { StoreInjector } from '../src/packages/storybook/withStore';
+import NL_PATHS from './static/gadm36_NLD_0.json';
 
 const selectedLocations = [
     {
@@ -55,6 +56,7 @@ storiesOf('Organisms|LocationSelector', module)
                     'Location selection'
                 )}
                 inputPlaceholder={text('Input placeholder', 'Location...')}
+                modalInputPlaceholder={text('Input placeholder inside the modal', '')}
                 noSuggestionsPlaceholder="noSuggestionsPlaceholder"
                 selectionPlaceholder="selectionPlaceholder"
                 doneLabel={text('Label for Done button', 'Done')}
@@ -67,6 +69,9 @@ storiesOf('Organisms|LocationSelector', module)
                 onRemoveAllLocations={() => console.log('onRemoveAllLocations was called')}
                 onLocationAutocompleteError={() =>
                     console.log('onLocationAutocompleteError was called')
+                }
+                defaultHighlight={
+                    boolean('Add default highlight area', true) ? NL_PATHS : undefined
                 }
             />
         );
@@ -119,7 +124,7 @@ storiesOf('Organisms|LocationSelector', module)
                 selectedLocations={store.get('selectedLocations')}
                 country={text('country', 'NL')}
                 language={text('Language', 'EN')}
-                initialMapAddress={text('Initial map address', '')}
+                initialMapAddress={text('Initial map address', 'Netherlands')}
                 radiusUnits={select('Radius units', ['km', 'mi'], 'km')}
                 renderRadiusLabel={(r) => `+ ${r} km`}
                 hasRadius={boolean('Has radius', true)}
@@ -136,6 +141,7 @@ storiesOf('Organisms|LocationSelector', module)
                     'Location selection'
                 )}
                 inputPlaceholder={text('Input placeholder', 'Location...')}
+                modalInputPlaceholder={text('Input placeholder inside the modal', '')}
                 noSuggestionsPlaceholder="noSuggestionsPlaceholder"
                 selectionPlaceholder={getPlaceholder()}
                 doneLabel={text('Label for Done button', 'Done')}
@@ -149,6 +155,9 @@ storiesOf('Organisms|LocationSelector', module)
                 }}
                 onLocationAutocompleteError={() =>
                     console.log('onLocationAutocompleteError was called')
+                }
+                defaultHighlight={
+                    boolean('Add default highlight area', true) ? NL_PATHS : undefined
                 }
             />
         );
@@ -200,7 +209,7 @@ storiesOf('Organisms|LocationSelector', module)
                 selectedLocations={store.get('selectedLocations')}
                 country={text('country', 'NL')}
                 language={text('Language', 'EN')}
-                initialMapAddress={text('Initial map address', '')}
+                initialMapAddress={text('Initial map address', 'Netherlands')}
                 radiusUnits={select('Radius units', ['km', 'mi'], 'km')}
                 renderRadiusLabel={(r) => `+ ${r} km`}
                 hasRadius={boolean('Has radius', true)}
@@ -217,6 +226,7 @@ storiesOf('Organisms|LocationSelector', module)
                     'Location selection'
                 )}
                 inputPlaceholder={text('Input placeholder', 'Location...')}
+                modalInputPlaceholder={text('Input placeholder inside the modal', '')}
                 noSuggestionsPlaceholder="noSuggestionsPlaceholder"
                 selectionPlaceholder={getPlaceholder()}
                 doneLabel={text('Label for Done button', 'Done')}
@@ -230,6 +240,9 @@ storiesOf('Organisms|LocationSelector', module)
                 }}
                 onLocationAutocompleteError={() =>
                     console.log('onLocationAutocompleteError was called')
+                }
+                defaultHighlight={
+                    boolean('Add default highlight area', true) ? NL_PATHS : undefined
                 }
             />
         );